default the test buckets dir to the canonicalized temp dir
Previously we were using the XDG data dir to avoid symlinks, but there's
no particular guarantee that that's not going to be a symlink too. Using
the temp dir by default is also slightly nicer for a couple reasons:
It's sometimes faster (an in-memory tempfs on e.g. Arch), and it makes
overriding `$TMPDIR` or `%TMP%` sufficient to control where tests put
temp files, without needing to override `UV_INTERNAL__TEST_DIR` too.